Move createTreeItemsWithErrorHandling into parent
