+++ ui/index.d.ts
@@ -263,28 +263,26 @@ export declare abstract class AzExtParentTreeItem<TRoot = ISubscriptionRoot> ext
     //#endregion
 
     /**
+     * Used to ensure a single invalid object does not prevent display of other valid objects
+     * Invalid objects will be shown with the error and the object's name. If the name cannot be determined for any invalid objects, a TreeItem will be added to the end with a generic label like "Some items cannot be displayed"
+     * @param sourceArray The collection of source objects before converting to TreeItems
+     * @param invalidContextValue The context value to use for invalid source objects
+     * @param createTreeItem A function that converts a source object to a TreeItem. Return undefined if you want this object to be skipped.
+     * @param getLabelOnError A minimal function that gets the label to display for an invalid source object
+     */
+    createTreeItemsWithErrorHandling<TSource, TTreeItem>(
+        sourceArray: TSource[],
+        invalidContextValue: string,
+        createTreeItem: (source: TSource) => AzExtTreeItem<TTreeItem> | undefined | Promise<AzExtTreeItem<TTreeItem> | undefined>,
+        getLabelOnError: (source: TSource) => string | undefined | Promise<string | undefined>): Promise<AzExtTreeItem<TTreeItem>[]>;
+
+    /**
      * This class wraps createChildImpl and ensures the tree is updated correctly when an item is created
      */
     createChild(userOptions?: any): Promise<AzExtTreeItem<TRoot>>;
 
     getCachedChildren(): Promise<AzExtTreeItem<TRoot>[]>;
 }
-
-/**
- * Used to ensure a single invalid object does not prevent display of other valid objects
- * Invalid objects will be shown with the error and the object's name. If the name cannot be determined for any invalid objects, a TreeItem will be added to the end with a generic label like "Some items cannot be displayed"
- * @param treeItem The parent tree item
- * @param sourceArray The collection of source objects before converting to TreeItems
- * @param invalidContextValue The context value to use for invalid source objects
- * @param createTreeItem A function that converts a source object to a TreeItem. Return undefined if you want this object to be skipped.
- * @param getLabelOnError A minimal function that gets the label to display for an invalid source object
- */
-export declare function createTreeItemsWithErrorHandling<TSource, TTreeItem>(
-    treeItem: AzExtParentTreeItem<TTreeItem>,
-    sourceArray: TSource[],
-    invalidContextValue: string,
-    createTreeItem: (source: TSource) => AzExtTreeItem<TTreeItem> | undefined | Promise<AzExtTreeItem<TTreeItem> | undefined>,
-    getLabelOnError: (source: TSource) => string | undefined | Promise<string | undefined>): Promise<AzExtTreeItem<TTreeItem>[]>;
 
 export declare class UserCancelledError extends Error { }
