+++ ui/src/treeDataProvider/AzExtParentTreeItem.ts
@@ -3,6 +3,7 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+import * as path from 'path';
 import { TreeItemCollapsibleState } from 'vscode';
 import { IAzureQuickPickItem, IAzureQuickPickOptions, ISubscriptionRoot } from '../..';
 import * as types from '../..';
@@ -144,6 +145,46 @@ export abstract class AzExtParentTreeItem<TRoot = ISubscriptionRoot> extends AzE
         }
     }
 
+    public async createTreeItemsWithErrorHandling<TSource, TTreeItem>(
+        sourceArray: TSource[],
+        invalidContextValue: string,
+        createTreeItem: (source: TSource) => AzExtTreeItem<TTreeItem> | undefined | Promise<AzExtTreeItem<TTreeItem> | undefined>,
+        getLabelOnError: (source: TSource) => string | undefined | Promise<string | undefined>): Promise<AzExtTreeItem<TTreeItem>[]> {
+
+        const treeItems: AzExtTreeItem<TTreeItem>[] = [];
+        // tslint:disable-next-line:no-any
+        let unknownError: any;
+        await Promise.all(sourceArray.map(async (source: TSource) => {
+            try {
+                const item: AzExtTreeItem<TTreeItem> | undefined = await createTreeItem(source);
+                if (item) {
+                    treeItems.push(item);
+                }
+            } catch (error) {
+                let name: string | undefined;
+                try {
+                    name = await getLabelOnError(source);
+                } catch {
+                    // ignore
+                }
+
+                if (name) {
+                    treeItems.push(new InvalidTreeItem<TTreeItem>(this, name, error, invalidContextValue));
+                } else if (error && !unknownError) {
+                    unknownError = error;
+                }
+            }
+        }));
+
+        if (unknownError) {
+            // Display a generic error if there are any unknown items. Only the first error will be displayed
+            const message: string = localize('cantShowItems', 'Some items could not be displayed');
+            treeItems.push(new InvalidTreeItem<TTreeItem>(this, message, unknownError, invalidContextValue, ''));
+        }
+
+        return treeItems;
+    }
+
     private async loadMoreChildrenInternal(): Promise<void> {
         if (this._clearCache) {
             // Just in case implementers of `loadMoreChildrenImpl` re-use the same child node, we want to clear those caches as well
@@ -203,3 +244,38 @@ export abstract class AzExtParentTreeItem<TRoot = ISubscriptionRoot> extends AzE
 }
 
 type GetTreeItemFunction<TRoot> = () => Promise<AzExtTreeItem<TRoot>>;
+
+class InvalidTreeItem<TRoot> extends AzExtParentTreeItem<TRoot> {
+    public readonly contextValue: string;
+    public readonly label: string;
+    public readonly description: string;
+
+    // tslint:disable-next-line:no-any
+    private _error: any;
+
+    // tslint:disable-next-line:no-any
+    constructor(parent: AzExtParentTreeItem<TRoot>, label: string, error: any, contextValue: string, description: string = localize('invalid', 'Invalid')) {
+        super(parent);
+        this.label = label;
+        this._error = error;
+        this.contextValue = contextValue;
+        this.description = description;
+    }
+
+    public get iconPath(): string {
+        return path.join(__filename, '..', '..', '..', '..', 'resources', 'warning.svg');
+    }
+
+    public async loadMoreChildrenImpl(): Promise<AzExtTreeItem<TRoot>[]> {
+        throw this._error;
+    }
+
+    public hasMoreChildrenImpl(): boolean {
+        return false;
+    }
+
+    public isAncestorOfImpl(): boolean {
+        // never display invalid items in tree picker
+        return false;
+    }
+}
